Moved data structs out of connection classes
