+++ src/main/java/ru/spbau/mit/TorrentTrackerConnection.java
@@ -15,99 +15,6 @@ public class TorrentTrackerConnection extends Connection {
     public static final int ACTION_SOURCES = 3;
     public static final int ACTION_UPDATE = 4;
 
-    public static final class FileEntry {
-        public static FileEntry readFrom(DataInputStream dis, boolean hasId) throws IOException {
-            if (hasId) {
-                return new FileEntry(
-                        dis.readInt(),
-                        dis.readUTF(),
-                        dis.readLong()
-                );
-            } else {
-                return new FileEntry(
-                        dis.readUTF(),
-                        dis.readLong()
-                );
-            }
-        }
-
-        private boolean hasId;
-        private int id;
-        private String name;
-        private long size;
-
-        public FileEntry(int id, String name, long size) {
-            this.hasId = true;
-            this.id = id;
-            this.name = name;
-            this.size = size;
-        }
-
-        public FileEntry(String name, long size) {
-            this.hasId = false;
-            this.name = name;
-            this.size = size;
-        }
-
-        public boolean hasId() {
-            return hasId;
-        }
-        public int getId() {
-            return id;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public long getSize() {
-            return size;
-        }
-
-        public void setId(int id) {
-            this.hasId = true;
-            this.id = id;
-        }
-
-        public void writeTo(DataOutputStream dos) throws IOException {
-            if (hasId) {
-                dos.writeInt(id);
-            }
-            dos.writeUTF(name);
-            dos.writeLong(size);
-        }
-    }
-
-    public static class ClientInfo {
-        public static ClientInfo readFrom(DataInputStream dis) throws IOException {
-            return new ClientInfo(
-                    dis.readUnsignedShort(),
-                    readList(DataInputStream::readInt, dis)
-            );
-        }
-
-        private int port;
-        private List<Integer> ids;
-
-        public ClientInfo(int port, List<Integer> ids) {
-            this.port = port;
-            this.ids = ids;
-        }
-
-        public int getPort() {
-            return port;
-        }
-
-        public List<Integer> getIds() {
-            return ids;
-        }
-
-        public void writeTo(DataOutputStream dos) throws IOException {
-            dos.writeShort(port);
-            writeList(ids, DataOutputStream::writeInt, dos);
-        }
-    };
-
     public TorrentTrackerConnection(Socket socket) throws IOException {
         super(socket);
     }
